==> 01-intro.Rmd <==
# Introduction {#intro}

**Placeholder for introduction text. I'll write this last along with the Conclusion.**




This product uses the **bookdown** package [@R-bookdown], which was built on top of R Markdown and **knitr** [@xie2015].
